Modify send feature to use shared slice + selectors
