+++ apps/minifront/src/state/send/helpers.ts
@@ -1,0 +1,26 @@
+import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
+import { BalancesResponse } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1/view_pb';
+import { getDisplay } from '@penumbra-zone/getters/metadata';
+import { getMetadata } from '@penumbra-zone/getters/value-view';
+import { assetPatterns } from '@penumbra-zone/types/assets';
+import { ZQueryState } from '@penumbra-zone/zquery';
+import { isKnown } from '../helpers';
+
+const nonTransferableAssetPatterns = [
+  assetPatterns.proposalNft,
+  assetPatterns.auctionNft,
+  assetPatterns.lpNft,
+];
+
+const isTransferable = (metadata: Metadata) =>
+  nonTransferableAssetPatterns.every(pattern => !pattern.matches(getDisplay(metadata)));
+
+export const transferableBalancesResponsesSelector = (
+  zQueryState: ZQueryState<BalancesResponse[]>,
+) => ({
+  loading: zQueryState.loading,
+  error: zQueryState.error,
+  data: zQueryState.data?.filter(
+    balance => isKnown(balance) && isTransferable(getMetadata(balance.balanceView)),
+  ),
+});
