+++ apps/minifront/src/state/swap/helpers.ts
@@ -21,7 +21,7 @@ import { BalancesResponse } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumb
 import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
 import { getDisplay } from '@penumbra-zone/getters/metadata';
 import { ZQueryState } from '@penumbra-zone/zquery';
-import { getValueViewCaseFromBalancesResponse } from '@penumbra-zone/getters/balances-response';
+import { isKnown } from '../helpers';
 
 export const sendSimulateTradeRequest = ({
   assetIn,
@@ -95,9 +95,6 @@ const nonSwappableAssetPatterns = [
 
 const isSwappable = (metadata: Metadata) =>
   nonSwappableAssetPatterns.every(pattern => !pattern.matches(getDisplay(metadata)));
-
-const isKnown = (balancesResponse: BalancesResponse) =>
-  getValueViewCaseFromBalancesResponse.optional()(balancesResponse) === 'knownAssetId';
 
 export const swappableBalancesResponsesSelector = (
   zQueryState: ZQueryState<BalancesResponse[]>,
